create windows in separate function; flush key buffer where needed
The curses windows are now created in create_windows which can
later be called from the signal handler for screen resizes.
This means it is possible to change the screen size while
the program is running, but it's still necessary to perform
an action which redraws the screen (such as entering the menu)
in order to make it work. Implementing a full automatic refresh
during resizing the window doesn't seem worth the effort.

Use flushinp() (flush keyboard input buffer) where appropriate,
e.g. when scrolling (scrolling may be slow due to database
fetches, and once the user stops pressing the Page Up/Dwn key
there should be no further scrolling) or when waiting for
operations that take some time to finish, like ADIF export
or up/downloading LoTW data to prevent that an user's impatient
key strokes cause him to miss the display with the results,
but jumps right back to the menu.